+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Breaking: Changed return types of methods from `Response` to `Dict[str,Any]`
+
 ## [2.1.2] - 2023-08-07
 ### Fixed
 - Set minimum typing-extensions version instead of exact version.