+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-- Ignore `use_validation_api` in `submit_job` method, deprecated and will be removed
+- Ignore `use_validation_api` in `submit_job`, method deprecated and will be removed
+- Ignore `use_validation_api` and `sid_server` in `Utilities.validate_id_params`, methods deprecated and will be removed
 ### Added
 - Add git commit hook with pre-commit